requirements.txt
Python 3.8.6rc1

colorama==0.4.4
Pillow==8.2.0